[lm] Move unkprobe and dph to new plugins
-- New API registerSegmentListener(seg, f)
-- New API segSplit(line)
-- New API hmWrite(s)
-- Fixes bug unregistering listeners
